#41583 3.2. Конвертер валюты. Добавил тестов и возможность конвертировать дробное число валюты, убрал избыточное приведение типов

+++ chapter_001/src/main/java/ru/job4j/calculator/Converter.java
@@ -37,7 +37,7 @@ public class Converter {
      * @return - количество рублей
      */
     public double euroToRouble(double euro) {
-        return euro * (double) EURO_RATE;
+        return euro * EURO_RATE;
     }
 
     /**
@@ -47,6 +47,6 @@ public class Converter {
      * @return - количество рублей
      */
     public double dollarToRouble(double dollar) {
-        return dollar * (double) DOLLAR_RATE;
+        return dollar * DOLLAR_RATE;
     }
 }
